Refactor native video decoder for code reuse
